requirements.txt
ffmpeg-python
pyqt6
